backend/requirements.txt
# Core
pytz==2020.1
Django==3.1.3
gunicorn==20.0.4
newrelic==5.14.1.144
django-dotenv==1.4.2
django-environ
django-3-jet==1.0.8
celery==4.4.7
django-webpack-loader
django-js-reverse
django-import-export
python-decouple
psycopg2
brotlipy
django-log-request-id
dj-database-url
whitenoise
psutil
django-naomi

# For the persistence stores
redis==3.5.3
psycopg2

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1
django-summernote==0.8.11.6
django-celery-beat==2.0.0
django-activity-stream==0.9.0
django-money==1.1

# Rest apis
djangorestframework==3.11.0
Markdown==3.2.2
drf-yasg==1.17.1
django-filter==2.3.0
django-cors-headers==3.4.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
easy-thumbnails==2.7
django-auto-prefetching==0.1.8

# Social login
social-auth-core==3.3.3
social-auth-app-django==4.0.0

# Developer Tools
ipdb==0.13.3
ipython==7.16.1
flake8==3.8.3
yapf==0.30.0
sentry-sdk==0.16.0
django-inlinecss==0.3.0

# Testing
factory-boy==2.12.0
django-nose==1.4.6
nose-progressive==1.5.2
coverage==5.1

# Static and Media Storage
django-storages==1.9.1
boto3==1.14.16
